Old file import is working. Except with name containing a '. Main window now uses a MemberPanel, but still not using the MVC pattern. Minor correction on Member to allow reflective calls on all methods.

+++ src/org/clubrockisen/entities/Party.java
@@ -248,7 +248,7 @@ public class Party extends Entity implements Cloneable, java.lang.Comparable<Par
 	 */
 	@Comparable
 	public Long getDate () {
-		return date == null ? Integer.valueOf(0) : date;
+		return date == null ? Long.valueOf(0) : date;
 	}
 	
 	/**
